+++ NEWS.md
@@ -47,6 +47,8 @@
   this requirement. The implementation differs from Kevin Ushey's implementation 
   in Rcpp, which IMHO tries to do too much. 
 
+* New sugar function `filter` #140. 
+  
 # Rcpp11 3.1.0.1
 
 * `Rcpp11` is now compatible with the gcc 4.6.3 as shipped by Rtools. Therefore